完善仓库管理模块，新增采购渠道

+++ AdminEAP-web/src/main/java/com/radish/master/entity/Stock.java
@@ -26,7 +26,7 @@ public class Stock extends BaseEntity {
 
     @Header(name = "库存量")
     @Column(name = "stock_num")
-    private Integer stock_num;
+    private Double stock_num;
 
     @Header(name = "入库人员ID")
     @Column(name = "storage_person_id")
@@ -69,11 +69,11 @@ public class Stock extends BaseEntity {
         this.mat_id = mat_id;
     }
 
-    public Integer getStock_num() {
+    public Double getStock_num() {
         return stock_num;
     }
 
-    public void setStock_num(Integer stock_num) {
+    public void setStock_num(Double stock_num) {
         this.stock_num = stock_num;
     }
